+++ packages/@webex/plugin-meetings/src/meeting/index.ts
@@ -116,17 +116,17 @@ import ControlsOptionsManager from '../controls-options-manager';
 
 const {isBrowser} = BrowserDetection();
 
-const logRequest = (request: any, {header = '', success = '', failure = ''}) => {
-  LoggerProxy.logger.info(header);
+const logRequest = (request: any, {logText = ''}) => {
+  LoggerProxy.logger.info(`${logText} - sending request`);
 
   return request
     .then((arg) => {
-      LoggerProxy.logger.info(success);
+      LoggerProxy.logger.info(`${logText} - has been successfully sent`);
 
       return arg;
     })
     .catch((error) => {
-      LoggerProxy.logger.error(failure, error);
+      LoggerProxy.logger.error(`${logText} - has failed: `, error);
       throw error;
     });
 };
@@ -3490,8 +3490,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no audio control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#muteAudio -->';
-
     // First, stop sending the local audio media
     return logRequest(
       this.audio
@@ -3515,9 +3513,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} muting audio`,
-        success: `${LOG_HEADER} muted audio successfully`,
-        failure: `${LOG_HEADER} muting audio failed, `,
+        logText: `Meeting:index#muteAudio --> correlationId=${this.correlationId} muting audio`,
       }
     );
   }
@@ -3543,8 +3539,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no audio control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#unmuteAudio -->';
-
     // First, send the control to unmute the participant on the server
     return logRequest(
       this.audio
@@ -3568,9 +3562,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} unmuting audio`,
-        success: `${LOG_HEADER} unmuted audio successfully`,
-        failure: `${LOG_HEADER} unmuting audio failed, `,
+        logText: `Meeting:index#unmuteAudio --> correlationId=${this.correlationId} unmuting audio`,
       }
     );
   }
@@ -3596,8 +3588,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no video control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#muteVideo -->';
-
     return logRequest(
       this.video
         .handleClientRequest(this, true)
@@ -3620,9 +3610,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} muting video`,
-        success: `${LOG_HEADER} muted video successfully`,
-        failure: `${LOG_HEADER} muting video failed, `,
+        logText: `Meeting:index#muteVideo --> correlationId=${this.correlationId} muting video`,
       }
     );
   }
@@ -3648,8 +3636,6 @@ export default class Meeting extends StatelessWebexPlugin {
       return Promise.reject(new ParameterError('no audio control associated to the meeting'));
     }
 
-    const LOG_HEADER = 'Meeting:index#unmuteVideo -->';
-
     return logRequest(
       this.video
         .handleClientRequest(this, false)
@@ -3672,9 +3658,7 @@ export default class Meeting extends StatelessWebexPlugin {
           throw error;
         }),
       {
-        header: `${LOG_HEADER} unmuting video`,
-        success: `${LOG_HEADER} unmuted video successfully`,
-        failure: `${LOG_HEADER} unmuting video failed, `,
+        logText: `Meeting:index#unmuteVideo --> correlationId=${this.correlationId} unmuting video`,
       }
     );
   }
@@ -4799,7 +4783,7 @@ export default class Meeting extends StatelessWebexPlugin {
     this.mediaProperties.webrtcMediaConnection.on(Event.ROAP_FAILURE, this.handleRoapFailure);
 
     this.mediaProperties.webrtcMediaConnection.on(Event.ROAP_MESSAGE_TO_SEND, (event) => {
-      const LOG_HEADER = 'Meeting:index#setupMediaConnectionListeners.ROAP_MESSAGE_TO_SEND -->';
+      const LOG_HEADER = `Meeting:index#setupMediaConnectionListeners.ROAP_MESSAGE_TO_SEND --> correlationId=${this.correlationId}`;
 
       switch (event.roapMessage.messageType) {
         case 'OK':
@@ -4815,9 +4799,7 @@ export default class Meeting extends StatelessWebexPlugin {
               correlationId: this.correlationId,
             }),
             {
-              header: `${LOG_HEADER} Send Roap OK`,
-              success: `${LOG_HEADER} Successfully send roap OK`,
-              failure: `${LOG_HEADER} Error joining the call on send roap OK, `,
+              logText: `${LOG_HEADER} Roap OK`,
             }
           );
           break;
@@ -4837,9 +4819,7 @@ export default class Meeting extends StatelessWebexPlugin {
               reconnect: this.reconnectionManager.isReconnectInProgress(),
             }),
             {
-              header: `${LOG_HEADER} Send Roap Offer`,
-              success: `${LOG_HEADER} Successfully send roap offer`,
-              failure: `${LOG_HEADER} Error joining the call on send roap offer, `,
+              logText: `${LOG_HEADER} Roap Offer`,
             }
           );
           break;
@@ -4858,9 +4838,7 @@ export default class Meeting extends StatelessWebexPlugin {
               correlationId: this.correlationId,
             }),
             {
-              header: `${LOG_HEADER} Send Roap Answer.`,
-              success: `${LOG_HEADER} Successfully send roap answer`,
-              failure: `${LOG_HEADER} Error joining the call on send roap answer, `,
+              logText: `${LOG_HEADER} Roap Answer`,
             }
           ).catch((error) => {
             const metricName = BEHAVIORAL_METRICS.ROAP_ANSWER_FAILURE;
@@ -4897,9 +4875,7 @@ export default class Meeting extends StatelessWebexPlugin {
               correlationId: this.correlationId,
             }),
             {
-              header: `${LOG_HEADER} Send Roap Error.`,
-              success: `${LOG_HEADER} Successfully send roap error`,
-              failure: `${LOG_HEADER} Failed to send roap error, `,
+              logText: `${LOG_HEADER} Roap Error (${event.roapMessage.errorType})`,
             }
           );
           break;
@@ -4993,7 +4969,7 @@ export default class Meeting extends StatelessWebexPlugin {
       };
 
       LoggerProxy.logger.info(
-        `Meeting:index#setupMediaConnectionListeners --> connection state changed to ${event.state}`
+        `Meeting:index#setupMediaConnectionListeners --> correlationId=${this.correlationId} connection state changed to ${event.state}`
       );
       switch (event.state) {
         case ConnectionState.Connecting: